requirements.txt
aiogram>=3.0
gspread
oauth2client
python-dotenv
pytz
pandas
openpyxl
